renamed user grains namespaces
